+++ ecmwf_models/grid.py
@@ -28,12 +28,13 @@ Common grid definitions for ECMWF model reanalysis products (regular gridded)
 import numpy as np
 from pygeogrids.grids import BasicGrid
 
+
 def get_grid_resolution(lats, lons):
     lats = np.unique(lats)
     lons = np.unique(lons)
     lats_res, lons_res = [], []
 
-    for i,j in zip(lats[:-1], lats[1:]):
+    for i, j in zip(lats[:-1], lats[1:]):
         lats_res.append(np.abs(np.abs(j)-np.abs(i)))
     lats_res = np.round(np.array(lats_res), 3)
     if not all(lats_res == lats_res[0]):
@@ -41,7 +42,7 @@ def get_grid_resolution(lats, lons):
     else:
         lat_res = lats_res[0]
 
-    for i,j in zip(lons[:-1], lons[1:]):
+    for i, j in zip(lons[:-1], lons[1:]):
         lons_res.append(np.abs(np.abs(j)-np.abs(i)))
     lons_res = np.round(np.array(lons_res), 3)
     if not all(lons_res == lons_res[0]):
@@ -49,6 +50,7 @@ def get_grid_resolution(lats, lons):
     else:
         lon_res = lons_res[0]
     return lat_res, lon_res
+
 
 def ERA_RegularImgGrid(res_lat=0.25, res_lon=0.25):
     '''
@@ -76,14 +78,14 @@ def ERA_RegularImgGrid(res_lat=0.25, res_lon=0.25):
 
     return BasicGrid(lon.flatten(), lat.flatten()).to_cell_grid(cellsize=5.)
 
+
 def ERA_IrregularImgGrid(lons, lats):
-        lons_gt_180 = np.where(lons > 180.0)
-        lons[lons_gt_180] = lons[lons_gt_180] - 360
-        return BasicGrid(lons.flatten(), lats.flatten()).to_cell_grid(cellsize=5.)
+    lons_gt_180 = np.where(lons > 180.0)
+    lons[lons_gt_180] = lons[lons_gt_180] - 360
+    return BasicGrid(lons.flatten(), lats.flatten()).to_cell_grid(cellsize=5.)
 
 
 def ERA_LandGrid():
-    #TODO: add function to generate TS from land points only,
+    # TODO: add function to generate TS from land points only,
     # use land mask (param: 172) to detect land points
     raise NotImplementedError
-