+++ CHANGES.rst
@@ -9,6 +9,8 @@ Version 0.X
 - Fix bugs, update command line interfaces, update dependencies
 - Separate download programs for ERA5 and ERA Interim
 - Change the ERA5 download api to use cdsapi instead of ecmwf api
+- Update package structure to better separate between the ERA products
+- Add look-up-table file for more flexibility in variable names passed by user
 - Update readme
 
 Version 0.4